+++ src/prefect/utilities/__init__.py
@@ -1,4 +1,5 @@
 import prefect.utilities.logging
+import prefect.utilities.backend
 import prefect.utilities.collections
 import prefect.utilities.configuration
 import prefect.utilities.datetimes